* master: (33 commits)
  Fixing the client build
  URL loader is now updated to Material-UI
  Two more built-in models, and some cleanup
  Added more built-in designs, and resurrected the spinner
  SimpleMeshJson import now supported (without panels)
  Now supporting all shapes
  Alerts are working again, with Material-UI
  JSweet now opens the URL from the query parameter
  WIP on new Material-UI
  WIP Groups are working
  Supporting all field symmetries correctly
  Now supporting editNumber correctly
  LinePlaneIntersection works now
  Enabled more files to parse
  AdjustSelectionByClass now working
  Missing file from prior commit
  MapToColor command works, mostly.
  Tool configs now working
  Default bookmark tool is working
  Panels are working
  ...

+++ core/src/main/java/com/vzome/core/construction/Color.java
@@ -128,9 +128,9 @@ public class Color
     public static Color parseWebColor( String colorStr )
     {
         return new Color(
-                Integer.valueOf( colorStr.substring( 1, 3 ), 16 ),
-                Integer.valueOf( colorStr.substring( 3, 5 ), 16 ),
-                Integer.valueOf( colorStr.substring( 5, 7 ), 16 ) );
+                Integer.valueOf( colorStr.substring( 1, 3 ), 16 ).intValue(),
+                Integer.valueOf( colorStr.substring( 3, 5 ), 16 ).intValue(),
+                Integer.valueOf( colorStr.substring( 5, 7 ), 16 ).intValue() );
     }
 
     public int getRed()
